Okay, because we are adding a sound on the homepage. There is no need for a AddSound page so It is now uploading sound page. I have also gotten rid of the sound id detail and the file_url.
