+++ src/layout/Navbar.jsx
@@ -21,6 +21,15 @@ export default function Navbar() {
                         <span className="navbar-toggler-icon"></span>
                     </button>
 
+                    <div className="collapse navbar-collapse" id="navbarSupportedContent">
+                        <ul className="navbar-nav me-auto">
+                            <li className="nav-item">
+                                <Link className="btn btn-outline-light" to="/uploadsounds">
+                                    Upload Sound
+                                </Link>
+                            </li>
+                        </ul>
+
                     <div className="ms-auto">
                         <Link to="/signup" className="btn btn-outline-light me-2">
                             Sign Up
@@ -29,8 +38,10 @@ export default function Navbar() {
                             Login
                         </Link>
                     </div>
+                    </div>
                 </div>
             </nav>
         </div>
+        
     );
 }
